feat: change the code structure
