+++ packages/runtime/remote-component/src/apis/loadComponent.ts
@@ -1,0 +1,58 @@
+import { assert, isAbsolute } from '@garfish/utils';
+import { Actuator } from '../actuator';
+import {
+  loader,
+  LOADING,
+  purifyOptions,
+  ComponentInfo,
+  cacheComponents,
+} from '../common';
+
+export function loadComponent(
+  options: ComponentInfo | string,
+): Promise<Record<string, any> | null> {
+  const info = purifyOptions(options);
+  const { url, env, cache, version, error, adapter } = info;
+
+  assert(url, 'Missing url for loading micro component');
+  assert(
+    isAbsolute(url),
+    'The loading of the micro component must be an absolute path.',
+  );
+
+  // `1.0@https://xx.js`
+  // `latest@https://xx.js`
+  const urlWithVersion = `${version || 'latest'}@${url}`;
+
+  const asyncLoadProcess = async () => {
+    let result = null;
+    try {
+      const component = cacheComponents[urlWithVersion];
+      if (cache && component) {
+        result = component;
+      } else {
+        const data = await loader.loadComponent(url);
+        const actuator = new Actuator(data.resourceManager, env);
+        let exports = actuator.execScript().exports;
+        if (typeof adapter === 'function') {
+          exports = adapter(exports);
+        }
+        result = exports;
+        cacheComponents[urlWithVersion] = result;
+      }
+    } catch (err) {
+      if (typeof error === 'function') {
+        result = error(err);
+      } else {
+        throw err;
+      }
+    } finally {
+      LOADING[urlWithVersion] = null;
+    }
+    return result;
+  };
+  if (!LOADING[urlWithVersion]) {
+    LOADING[urlWithVersion] = asyncLoadProcess();
+  }
+  return LOADING[urlWithVersion];
+}
