+++ packages/runtime/remote-component/src/apis/preload.ts
@@ -1,0 +1,19 @@
+import { assert, isAbsolute } from '@garfish/utils';
+import { loader, PRE_STORED_RESOURCES } from '../common';
+
+// Preload the static resources of the component, so that the component can be loaded synchronously
+export function preload(urls: string | Array<string>) {
+  if (!Array.isArray(urls)) urls = [urls];
+
+  return Promise.all(
+    urls.map((url) => {
+      assert(
+        isAbsolute(url || ''),
+        'The loading of the micro component must be an absolute path.',
+      );
+      return loader.loadComponent(url).then((data) => {
+        PRE_STORED_RESOURCES[url] = data.resourceManager;
+      });
+    }),
+  );
+}
